cdr/requirements.txt
faker
paramiko
confluent-kafka
cassandra-driver==3.25.0
